+++ app/index.js
@@ -3,12 +3,11 @@ import { staticPlugin } from '@elysiajs/static'
 import { nullish__none_, tup } from 'ctx-core/all'
 import { be_memo_pair_, be_sig_triple_, rmemo__wait } from 'ctx-core/rmemo'
 import { Elysia } from 'elysia'
-import { dirname, join, relative } from 'path'
+import { dirname, join } from 'path'
 import {
 	app__relative_path_,
 	app_ctx,
 	app_ctx__be_config,
-	app_path_,
 	browser__metafile$_,
 	browser__metafile_path_,
 	browser_path_,
@@ -61,7 +60,11 @@ export const [
 ] = be_memo_pair_(ctx=>
 	join(dirname(server_entry__output__path_(ctx)), 'index.js'),
 { ...app_ctx__be_config, id: 'server_entry__output__link__path' })
-export async function app__new() {
+/**
+ * @param {Elysia}[app]
+ * @returns {Promise<Elysia>}
+ */
+export async function app__attach(app) {
 	if (!await file_exists_(server__metafile_path_(app_ctx))) {
 		throw new Error(`${server__metafile_path_(app_ctx)} does not exist`)
 	}
@@ -73,11 +76,11 @@ export async function app__new() {
 		server__metafile$_(app_ctx),
 		neq_undefined)
 	await rmemo__wait(browser__metafile$_(app_ctx), neq_undefined)
-	const app =
-		new Elysia().use(staticPlugin({
-			assets: browser_path_(app_ctx),
-			prefix: '',
-		}))
+	app ??= new Elysia()
+	app.use(staticPlugin({
+		assets: browser_path_(app_ctx),
+		prefix: '',
+	}))
 	const { inputs } = server__metafile
 	for (let input_path in inputs) {
 		const middleware_ctx = middleware_ctx_()
@@ -92,12 +95,12 @@ export async function app__new() {
 	}
 	return app
 }
-export async function app__start() {
-	let app = app_(app_ctx)
-	if (app) {
-		await app.stop()
-	}
-	app = await app__new()
+/**
+ * @param {Elysia}[app]
+ * @returns {Promise<Elysia>}
+ */
+export async function app__start(app) {
+	app ??= await app__attach()
 	app__set(app_ctx, app)
 	const port = port_(app_ctx)
 	app.listen(port)